pyproject.toml
[project]
name = "resumeai-test1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "faiss-cpu>=1.13.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "pdfplumber>=0.11.8",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
]
